core/requirements.txt
flask
mediapipe
numpy
Pillow
gunicorn
